This closes #15. CRC32 calculation is also updated.

+++ FW/hw/inc/sm_hw_storage.h
@@ -112,7 +112,7 @@ private:
 
     SmHalAbstractSpi *mSpi;
 
-    uint8_t mCount;    ///< Number of elements in the storage
+    uint32_t mCount;    ///< Number of elements in the storage
 
     // SmHwPowerMgrIface
     void onSleep(void);
